refactor(registrar): extract WelcomeView from RegistrarPanel (−394 lines)
Extracts the inline welcome dashboard view (lines 1586-2020 of
RegistrarPanel.jsx) into a dedicated src/pages/registrar/views/WelcomeView.jsx
component, mirroring the QueueView.jsx extraction pattern (decomp 6a → 6b).

The welcome view was tightly coupled to ~34 parent scope items: translation
fn (t), language, theme, textColor, appointments, departmentStats,
dataSource, appointmentsLoading, filteredAppointments, services, activeTab,
historyDate, showCalendar, tempDateInput, loadAppointments, setShowWizard,
setWizardEditMode, setWizardInitialData, setShowPaymentManager,
setHistoryDate, setShowCalendar, setTempDateInput, setSearchParams,
navigate, setPaymentDialog, setPrintDialog, setContextMenu,
openRecordPreview, openRecordEditor, updateAppointmentStatus,
handleStartVisit, plus inline-defined helpers generateCSV, downloadCSV and
the memoized DataSourceIndicator component — all now passed as explicit
props.

Result: RegistrarPanel.jsx shrinks 2700 → 2306 lines (−394). The welcome
view's JSX, classNames, aria-labels, event handlers, and structure are
preserved verbatim. Tests: 467/467 passing.
